Move test board setup to conftest
Moving the test board setup to conftest.py will save some duplicated code overall. Instead of having several files with an identical setup function, we simply have the single setup for all tests. To use it, we just have to include it in the test function parameters, and then use it as a tuple/list. In the future, splitting boardOne and boardTwo setups may be preferential.